Implemented DBM-in-memory

+++ dbm-persistence/Statements/GetInstrumentListStmt.cs
@@ -29,7 +29,7 @@ internal sealed class GetInstrumentListStmt : QueryDbStmtBase {
 
     protected override bool ProcessCurrentRow(NpgsqlDataReader reader) {
         var i = new InstrumentDto(
-            (ulong)reader.GetInt64(0),
+            reader.GetInt64(0),
             reader.GetString(1),
             reader.GetString(2),
             reader.GetString(3),
